refactor(terrarium): split engine_cli TUI helpers below size cap
engine_cli.py grew past the repo's 600-line soft cap on main (607
lines), which fails tests/unit/test_file_sizes.py for every fresh PR
against current main. Move the tab-strip / panel / channel-tab helpers
into a sibling engine_cli_tabs module and point the topology-refresh
test at its new home. Pure move, no behavior change.
